Removed dead code.  Initiated debug text at page load.
Jumping from the top of the page to the bottom causes the top item to be appended to the bottom approximately 100,000 times.

My next task is to recognize that a large  distance was traveled at once and to just rebuild the list instead of updating it one item at a time.