Fixed the UI for Firefox for real this time.
It turns out setting the width property of a table cell on Firefox causes it to disable automatic resizing.  On Chrome, setting the width to 0 causes the auto resizing to stay on.

All items are now centered on Firefox and Chrome, and the append button fills the remaining space in each row.